Split config.rs into concern submodules
config.rs had grown to fuse five separable concerns behind one 2000-line
file: keyring-store installation, dev-mode env detection and keyring
seeding, the export-settings value domain, atomic-write machinery, and a
stray RFC 3339 time helper that has nothing to do with config. Bundling
them made the config concern (load, discovery, active-library pointer,
reactive handle, config value types) hard to see through the noise.

Each concern moves to its own file. The keyring, dev, export, and
atomic-write modules stay private under config; config.rs re-exports their
public items, so every existing bae_core::config::* path is byte-identical
and no caller changes. The time helper's old path was simply wrong — it is
not config — so it moves to crate::util::time and its one bae-core caller
is updated outright.
