+++ bae-core/src/util/mod.rs
@@ -5,3 +5,4 @@ pub mod format;
 pub mod fs;
 pub mod http;
 pub mod session_cache;
+pub mod time;
